+++ packages/dev-server-core/src/server/createServer.ts
@@ -9,7 +9,7 @@ import net, { Server, Socket, ListenOptions } from 'net';
 import { DevServerCoreConfig } from '../DevServerCoreConfig';
 import { createMiddleware } from './createMiddleware';
 import { Logger } from '../logger/Logger';
-import { createPlugins } from './createPlugins';
+import { addPlugins } from './addPlugins';
 
 /**
  * A request handler that returns a 301 HTTP Redirect to the same location as the original
@@ -28,22 +28,18 @@ function httpsRedirect(req: IncomingMessage, res: ServerResponse) {
 export function createServer(cfg: DevServerCoreConfig, logger: Logger, fileWatcher: FSWatcher) {
   const app = new Koa();
 
-  const plugins = createPlugins(cfg);
-  if (!cfg.plugins) {
-    cfg.plugins = [];
-  }
-  cfg.plugins.push(...plugins);
+  addPlugins(cfg);
 
   // special case the legacy plugin, if it is given make sure the resolve module imports plugin
   // runs before the legacy plugin because it compiles away module syntax. ideally we have a
   // generic API for this, but we need to design that a bit more first
-  const indexOfLegacy = cfg.plugins.findIndex(p => p.name === 'legacy');
-  let indexOfResolve = cfg.plugins.findIndex(p => p.name === 'resolve-module-imports');
+  const indexOfLegacy = cfg.plugins!.findIndex(p => p.name === 'legacy');
+  let indexOfResolve = cfg.plugins!.findIndex(p => p.name === 'resolve-module-imports');
   if (indexOfLegacy !== -1 && indexOfResolve !== -1) {
-    const legacy = cfg.plugins.splice(indexOfLegacy, 1)[0];
+    const legacy = cfg.plugins!.splice(indexOfLegacy, 1)[0];
     // recompute after splicing
-    indexOfResolve = cfg.plugins.findIndex(p => p.name === 'resolve-module-imports');
-    cfg.plugins.splice(indexOfResolve, 1, cfg.plugins[indexOfResolve], legacy);
+    indexOfResolve = cfg.plugins!.findIndex(p => p.name === 'resolve-module-imports');
+    cfg.plugins!.splice(indexOfResolve, 1, cfg.plugins![indexOfResolve], legacy);
   }
 
   const middleware = createMiddleware(cfg, logger, fileWatcher);