+++ packages/dev-server-core/src/web-sockets/WebSocketsManager.ts
@@ -4,8 +4,10 @@ import { EventEmitter } from './EventEmitter';
 
 export const NAME_WEB_SOCKET_IMPORT = '/__web-dev-server__web-socket.js';
 
+type WebSocketData = { type: string } & Record<string, unknown>;
+
 export interface Events {
-  message: { webSocket: WebSocket; data: WebSocket.Data };
+  message: { webSocket: WebSocket; data: WebSocketData };
 }
 
 /**
@@ -14,7 +16,7 @@ export interface Events {
  */
 export class WebSocketsManager extends EventEmitter<Events> {
   public webSocketImport = NAME_WEB_SOCKET_IMPORT;
-  private webSocketServer: WebSocket.Server;
+  public webSocketServer: WebSocket.Server;
   private openSockets = new Set<WebSocket>();
 
   constructor(server: Server) {
@@ -27,8 +29,17 @@ export class WebSocketsManager extends EventEmitter<Events> {
         this.openSockets.delete(webSocket);
       });
 
-      webSocket.on('message', data => {
-        this.emit('message', { webSocket, data });
+      webSocket.on('message', rawData => {
+        try {
+          const data = JSON.parse(rawData.toString());
+          if (!data.type) {
+            throw new Error('Missing property "type".');
+          }
+          this.emit('message', { webSocket, data });
+        } catch (error) {
+          console.error('Failed to parse websocket event received from the browser: ', rawData);
+          console.error(error);
+        }
       });
     });
 
@@ -40,7 +51,7 @@ export class WebSocketsManager extends EventEmitter<Events> {
   }
 
   /**
-   * Imports the given path, executing the module as well as a default export if it executes a function.
+   * Imports the given path, executing the module as well as a default export if it exports a function.
    *
    * This is a built-in web socket message and will be handled automatically.
    *