+++ packages/dev-server/src/config/parseConfig.ts
@@ -71,20 +71,21 @@ export async function parseConfig(
     finalConfig.port = await getPortPromise({ port: 8000 });
   }
 
-  // map flags to plugin
-  if (finalConfig?.esbuildTarget) {
-    finalConfig.plugins!.push(esbuildPlugin(finalConfig.esbuildTarget));
-  }
-
   if (finalConfig.nodeResolve) {
     const userOptions = typeof config.nodeResolve === 'object' ? config.nodeResolve : undefined;
+    // do node resolve after user plugins, to allow user plugins to resolve imports
     finalConfig.plugins!.push(
       nodeResolvePlugin(finalConfig.rootDir!, config.preserveSymlinks, userOptions),
     );
   }
 
+  // map flags to plugin
+  if (finalConfig?.esbuildTarget) {
+    finalConfig.plugins!.unshift(esbuildPlugin(finalConfig.esbuildTarget));
+  }
+
   if (finalConfig.watch) {
-    finalConfig.plugins!.push(watchPlugin());
+    finalConfig.plugins!.unshift(watchPlugin());
   }
 
   return finalConfig;