+++ packages/test-runner-core/src/browser-launcher/BrowserLauncher.ts
@@ -19,6 +19,14 @@ export interface BrowserLauncher {
   type: string;
 
   /**
+   * Optional concurrency for this browser launcher only. Overwrites a globally
+   * configured concurrency option.
+   */
+  concurrency?: number;
+
+  __experimentalWindowFocus__?: boolean;
+
+  /**
    * One time startup for the browser launcher. Called when the test runner
    * starts. Use this for async initialization work.
    * @param config
@@ -47,6 +55,12 @@ export interface BrowserLauncher {
    * @param session
    */
   isActive(sessionId: string): boolean;
+
+  /**
+   * Returns the current browser URL for the test session. This is used for example to
+   * detect browser navigations.
+   */
+  getBrowserUrl(sessionId: string): string | undefined | Promise<string | undefined>;
 
   /**
    * Stops a single test session. There is no mandatory action to be taken here.