+++ packages/test-runner-core/src/runner/TestScheduler.ts
@@ -7,8 +7,6 @@ import { withTimeout } from '../utils/async';
 import { TestSessionTimeoutHandler } from './TestSessionTimeoutHandler';
 import { BrowserLauncher } from '../browser-launcher/BrowserLauncher';
 
-const TIMEOUT_MSG = 'Timeout starting the browser page.';
-
 export class TestScheduler {
   private config: TestRunnerCoreConfig;
   private sessions: TestSessionManager;
@@ -16,6 +14,7 @@ export class TestScheduler {
   private browsers: BrowserLauncher[];
   private finishedBrowsers = new Set<BrowserLauncher>();
   private stopPromises = new Set<Promise<unknown>>();
+  private browserStartTimeoutMsg: string;
 
   constructor(
     config: TestRunnerCoreConfig,
@@ -24,8 +23,13 @@ export class TestScheduler {
   ) {
     this.config = config;
     this.sessions = sessions;
-    this.browsers = browsers;
+    this.browsers = [...browsers].sort(
+      (a, b) => (a.__experimentalWindowFocus__ ? 1 : 0) - (b.__experimentalWindowFocus__ ? 1 : 0),
+    );
     this.timeoutHandler = new TestSessionTimeoutHandler(this.config, this.sessions, this);
+    this.browserStartTimeoutMsg =
+      `The browser was unable to create and start a test page after ${this.config.browserStartTimeout}ms. ` +
+      'You can increase this timeout with the browserStartTimeout option.';
 
     sessions.on('session-status-updated', async session => {
       if (session.status === SESSION_STATUS.TEST_STARTED) {
@@ -75,6 +79,10 @@ export class TestScheduler {
     let runningBrowsers = 0;
 
     for (const browser of this.browsers) {
+      if (runningBrowsers >= 1 && browser.__experimentalWindowFocus__) {
+        return;
+      }
+
       if (runningBrowsers >= this.config.concurrentBrowsers) {
         // do not boot up more than the allowed concurrent browsers
         return;
@@ -86,7 +94,15 @@ export class TestScheduler {
         runningBrowsers += 1;
 
         const runningCount = this.getRunningSessions(browser).length;
-        const runBudget = Math.max(0, this.config.concurrency - runningCount);
+        let maxBudget;
+
+        if (browser.__experimentalWindowFocus__) {
+          maxBudget = 1;
+        } else {
+          maxBudget = browser.concurrency ?? this.config.concurrency;
+        }
+
+        const runBudget = Math.max(0, maxBudget - runningCount);
         if (runBudget !== 0) {
           // we have budget to schedule new sessions for this browser
           const allScheduled = this.getScheduledSessions(browser);
@@ -108,7 +124,7 @@ export class TestScheduler {
     const timeoutId = setTimeout(() => {
       if (!browserStarted && !this.timeoutHandler.isStale(updatedSession)) {
         this.setSessionFailed(this.sessions.get(updatedSession.id)!, {
-          message: `The browser was unable to open the test page after ${this.config.browserStartTimeout}ms.`,
+          message: this.browserStartTimeoutMsg,
         });
       }
     }, this.config.browserStartTimeout!);
@@ -120,7 +136,7 @@ export class TestScheduler {
           updatedSession.id,
           createSessionUrl(this.config, updatedSession),
         ),
-        TIMEOUT_MSG,
+        this.browserStartTimeoutMsg,
         this.config.browserStartTimeout,
       );
 
@@ -130,7 +146,7 @@ export class TestScheduler {
       if (this.timeoutHandler.isStale(updatedSession)) {
         // something else has changed the test session, such as a the browser timeout
         // or a re-run in watch mode. in that was we just log the error
-        if (error.message !== TIMEOUT_MSG) {
+        if (error.message !== this.browserStartTimeoutMsg) {
           this.config.logger.error(error);
         }
       } else {