+++ packages/test-runner-core/src/server/TestRunnerServer.ts
@@ -3,14 +3,15 @@ import chokidar from 'chokidar';
 
 import { RunSessions, watchFilesMiddleware } from './middleware/watchFilesMiddleware';
 import { cacheMiddleware } from './middleware/cacheMiddleware';
-import { testRunnerApiMiddleware } from './middleware/testRunnerApiMiddleware';
 import { serveTestRunnerHtmlPlugin } from './plugins/serveTestRunnerHtmlPlugin';
 import { serveTestFrameworkPlugin } from './plugins/serveTestFrameworkPlugin';
+import { testRunnerApiPlugin } from './plugins/testRunnerApiPlugin';
 import { TestRunnerCoreConfig } from '../config/TestRunnerCoreConfig';
 import { TestSessionManager } from '../test-session/TestSessionManager';
 
 const CACHED_PATTERNS = [
   'node_modules/@web/test-runner-',
+  'node_modules/@esm-bundle/chai',
   'node_modules/mocha/',
   'node_modules/chai/',
 ];
@@ -35,17 +36,18 @@ export class TestRunnerServer {
       port: config.port,
       hostname: config.hostname,
       rootDir,
+      injectWebSocket: true,
 
       mimeTypes: config.mimeTypes,
       middleware: [
-        testRunnerApiMiddleware(sessions, rootDir, config, plugins),
         watchFilesMiddleware({ runSessions, sessions, rootDir, fileWatcher: this.fileWatcher }),
         cacheMiddleware(CACHED_PATTERNS, config.watch),
         ...(config.middleware || []),
       ],
 
       plugins: [
-        serveTestRunnerHtmlPlugin(config, testFiles, testFrameworkImport),
+        testRunnerApiPlugin(config, plugins, sessions),
+        serveTestRunnerHtmlPlugin(config, testFiles, sessions, testFrameworkImport),
         testFrameworkPlugin,
         ...(config.plugins || []),
       ].filter(isDefined) as Plugin[],