+++ packages/test-runner-core/src/server/plugins/serveTestFrameworkPlugin.ts
@@ -9,7 +9,10 @@ const REGEXP_SOURCE_MAP = /\/\/# sourceMappingURL=.*/;
 
 async function readFile(codePath: string) {
   if (!fs.existsSync(codePath)) {
-    throw new Error(`Could not find a test framework at ${codePath}`);
+    throw new Error(
+      `The test framework at ${codePath} could not be loaded. ` +
+        'Are your dependencies installed correctly? Is there a server plugin or middleware that interferes?',
+    );
   }
 
   return (await promisify(fs.readFile)(codePath, 'utf-8')).replace(REGEXP_SOURCE_MAP, '');
@@ -21,9 +24,8 @@ async function readFile(codePath: string) {
 export function serveTestFrameworkPlugin(testFramework: TestFramework) {
   const testFrameworkFilePath = path.resolve(testFramework.path);
   const testeFrameworkBrowserPath = testFrameworkFilePath.split(path.sep).join('/');
-  const testFrameworkImport = path.posix.join(
-    TEST_FRAMEWORK_IMPORT_ROOT,
-    testeFrameworkBrowserPath,
+  const testFrameworkImport = encodeURI(
+    path.posix.join(TEST_FRAMEWORK_IMPORT_ROOT, testeFrameworkBrowserPath),
   );
 
   const testFrameworkPlugin: Plugin = {
@@ -31,7 +33,7 @@ export function serveTestFrameworkPlugin(testFramework: TestFramework) {
 
     async serve(context) {
       if (context.path.startsWith(TEST_FRAMEWORK_IMPORT_ROOT)) {
-        const importPath = context.path.replace(TEST_FRAMEWORK_IMPORT_ROOT, '');
+        const importPath = decodeURI(context.path.replace(TEST_FRAMEWORK_IMPORT_ROOT, ''));
         let filePath = importPath.split('/').join(path.sep);
         // for posix the leading / will be stripped by path.join above
         if (path.sep === '/') {