+++ packages/test-runner-core/src/test-helpers.ts
@@ -3,7 +3,7 @@ import { getPortPromise } from 'portfinder';
 import path from 'path';
 import { TestRunner, TestRunnerCoreConfig } from './index';
 import { Logger } from './logger/Logger';
-import { TestResult, TestSuiteResult } from './test-session/TestSession';
+import { TestResult, TestSession, TestSuiteResult } from './test-session/TestSession';
 import { SESSION_STATUS } from './test-session/TestSessionStatus';
 import { TestRunnerGroupConfig } from './config/TestRunnerGroupConfig';
 
@@ -17,6 +17,9 @@ const logger: Logger = {
   },
 };
 
+const secondMs = 1000;
+const minuteMs = secondMs * 60;
+
 const defaultBaseConfig: Partial<TestRunnerCoreConfig> = {
   watch: false,
   rootDir: path.join(__dirname, '..', '..', '..'),
@@ -28,9 +31,9 @@ const defaultBaseConfig: Partial<TestRunnerCoreConfig> = {
   reporters: [],
   concurrentBrowsers: 2,
   concurrency: 10,
-  browserStartTimeout: 30000,
-  testsStartTimeout: 10000,
-  testsFinishTimeout: 20000,
+  browserStartTimeout: minuteMs / 2,
+  testsStartTimeout: secondMs * 20,
+  testsFinishTimeout: minuteMs * 2,
   browserLogs: true,
   logger,
 };
@@ -42,13 +45,17 @@ export async function runTests(
     allowFailure = false,
     reportErrors = true,
   }: { allowFailure?: boolean; reportErrors?: boolean } = {},
-): Promise<TestRunner> {
+): Promise<{ runner: TestRunner; sessions: TestSession[] }> {
   return new Promise(async (resolve, reject) => {
     const port = await getPortPromise({ port: 9000 + Math.floor(Math.random() * 1000) });
     const finalConfig = {
       port,
       ...defaultBaseConfig,
       ...config,
+      testFramework: {
+        ...defaultBaseConfig.testFramework,
+        ...config.testFramework,
+      },
     } as TestRunnerCoreConfig;
 
     const runner = new TestRunner(finalConfig, groupConfigs);
@@ -131,10 +138,8 @@ export async function runTests(
       }
 
       if (allowFailure || passed) {
-        resolve(runner);
+        resolve({ runner, sessions });
       } else {
-        console.error('Failed tests:');
-        console.error(sessions.filter(s => !s.passed));
         reject(new Error('Test run did not pass'));
       }
     });