+++ packages/test-runner-playwright/src/index.ts
@@ -10,6 +10,8 @@ export interface PlaywrightLauncherArgs {
   product?: ProductType;
   launchOptions?: LaunchOptions;
   createPage?: (args: { config: TestRunnerCoreConfig; browser: Browser }) => Promise<Page>;
+  __experimentalWindowFocus__?: boolean;
+  concurrency?: number;
 }
 
 export { PlaywrightLauncher };
@@ -22,5 +24,11 @@ export function playwrightLauncher(args: PlaywrightLauncherArgs = {}) {
     );
   }
 
-  return new PlaywrightLauncher(product, args.launchOptions ?? {}, args.createPage);
+  return new PlaywrightLauncher(
+    product,
+    args.launchOptions ?? {},
+    args.createPage,
+    !!args.__experimentalWindowFocus__,
+    args.concurrency,
+  );
 }