+++ packages/test-runner-playwright/test/playwrightLauncher.test.ts
@@ -1,27 +1,77 @@
-import { runTests } from '@web/test-runner-core/test-helpers';
-import { resolve } from 'path';
-
+import os from 'os';
+import { runIntegrationTests } from '../../../integration/test-runner';
 import { playwrightLauncher } from '../src/index';
 
-it('runs tests with playwright', async function () {
-  this.timeout(50000);
+describe('test-runner-playwright chromium', function testRunnerPlaywright() {
+  this.timeout(100000);
 
-  await runTests({
-    files: [
-      resolve(__dirname, 'fixtures', 'a.js'),
-      resolve(__dirname, 'fixtures', 'b.js'),
-      resolve(__dirname, 'fixtures', 'c.js'),
-      resolve(__dirname, 'fixtures', 'd.js'),
-      resolve(__dirname, 'fixtures', 'e.js'),
-      resolve(__dirname, 'fixtures', 'f.js'),
-      resolve(__dirname, 'fixtures', 'g.js'),
-    ],
-    browsers: [
-      playwrightLauncher({ product: 'chromium' }),
-      playwrightLauncher({ product: 'firefox' }),
-      playwrightLauncher({ product: 'webkit' }),
-    ],
-    concurrentBrowsers: 2,
-    concurrency: 6,
+  function createConfig() {
+    return { browsers: [playwrightLauncher({ product: 'chromium' })] };
+  }
+
+  runIntegrationTests(createConfig, {
+    basic: true,
+    groups: true,
+    parallel: true,
+    testFailure: true,
+    locationChanged: true,
   });
 });
+
+// we don't run all tests in the windows CI
+if (os.platform() !== 'win32') {
+  describe('test-runner-playwright webkit', function testRunnerPlaywright() {
+    this.timeout(100000);
+
+    function createConfig() {
+      return { browsers: [playwrightLauncher({ product: 'webkit' })] };
+    }
+
+    runIntegrationTests(createConfig, {
+      basic: true,
+      groups: true,
+      parallel: true,
+      testFailure: true,
+      locationChanged: true,
+    });
+  });
+
+  describe('test-runner-playwright firefox', function testRunnerPlaywright() {
+    this.timeout(100000);
+
+    function createConfig() {
+      return { browsers: [playwrightLauncher({ product: 'firefox' })] };
+    }
+
+    runIntegrationTests(createConfig, {
+      basic: true,
+      groups: true,
+      // firefox doesn't like parallel in the CI
+      parallel: false,
+      testFailure: true,
+      locationChanged: true,
+    });
+  });
+
+  describe('test-runner-playwright all', function testRunnerPlaywright() {
+    this.timeout(100000);
+
+    function createConfig() {
+      return {
+        browsers: [
+          playwrightLauncher({ product: 'chromium' }),
+          playwrightLauncher({ product: 'firefox' }),
+          ...(os.platform() !== 'win32' ? [playwrightLauncher({ product: 'webkit' })] : []),
+        ],
+      };
+    }
+
+    runIntegrationTests(createConfig, {
+      basic: true,
+      groups: true,
+      parallel: false,
+      testFailure: false,
+      locationChanged: false,
+    });
+  });
+}