+++ packages/test-runner-puppeteer/test/puppeteerLauncher.test.ts
@@ -1,25 +1,20 @@
-import { runTests } from '@web/test-runner-core/test-helpers';
-import { resolve } from 'path';
+import { runIntegrationTests } from '../../../integration/test-runner';
+import { puppeteerLauncher } from '../src/index';
 
-import { puppeteerLauncher } from '../src/puppeteerLauncher';
+describe('test-runner-puppeteer', function testRunnerPuppeteer() {
+  this.timeout(20000);
 
-it('runs tests with puppeteer', async function () {
-  this.timeout(50000);
+  function createConfig() {
+    return {
+      browsers: [puppeteerLauncher()],
+    };
+  }
 
-  await runTests({
-    files: [
-      resolve(__dirname, 'fixtures', 'a.js'),
-      resolve(__dirname, 'fixtures', 'b.js'),
-      resolve(__dirname, 'fixtures', 'c.js'),
-      resolve(__dirname, 'fixtures', 'd.js'),
-      resolve(__dirname, 'fixtures', 'e.js'),
-      resolve(__dirname, 'fixtures', 'f.js'),
-      resolve(__dirname, 'fixtures', 'g.js'),
-    ],
-    browsers: [puppeteerLauncher()],
-    // firefox doesn't work in our CI
-    // browsers: [puppeteerLauncher({ launchOptions: { product: 'firefox' } })],
-    concurrentBrowsers: 2,
-    concurrency: 6,
+  runIntegrationTests(createConfig, {
+    basic: true,
+    groups: true,
+    parallel: true,
+    testFailure: true,
+    locationChanged: true,
   });
 });