+++ packages/test-runner/demo/groups.config.mjs
@@ -1,4 +1,3 @@
-//@ts-check
 import { playwrightLauncher } from '@web/test-runner-playwright';
 
 export default /** @type {import('@web/test-runner').TestRunnerConfig} */ ({
@@ -7,13 +6,23 @@ export default /** @type {import('@web/test-runner').TestRunnerConfig} */ ({
 
   groups: [
     {
+      name: 'chromium-a',
+      files: 'demo/test/pass-*.test.js',
+      browsers: [playwrightLauncher({ product: 'chromium' })],
+    },
+    {
+      name: 'chromium-b',
+      files: 'demo/test/pass-*.test.js',
+      browsers: [playwrightLauncher({ product: 'chromium' })],
+    },
+    {
       name: 'firefox',
-      files: 'demo/test/pass-1.test.js',
+      files: 'demo/test/pass-*.test.js',
       browsers: [playwrightLauncher({ product: 'firefox' })],
     },
     {
       name: 'webkit',
-      files: 'demo/test/pass-2.test.js',
+      files: 'demo/test/pass-*.test.js',
       browsers: [playwrightLauncher({ product: 'webkit' })],
     },
   ],