+++ packages/dev-server-esbuild/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @web/dev-server-esbuild
 
+## 0.2.6
+
+### Patch Changes
+
+- 46a01fb: filter unsupported sourcemap warnings
+
+## 0.2.5
+
+### Patch Changes
+
+- 8111c2f: upgrade mdn-browser-compat-data to @mdn/browser-compat-data
+- 201ffbd: updated esbuild dependency
+
 ## 0.2.4
 
 ### Patch Changes