+++ packages/dev-server-legacy/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @web/dev-server-legacy
 
+## 0.1.6
+
+### Patch Changes
+
+- 7aabb52: only transform modules to systemjs
+- Updated dependencies [e8ebfcc]
+  - @web/dev-server-core@0.2.13
+
+## 0.1.5
+
+### Patch Changes
+
+- 7901081: Include URL parameters in inline scripts cache
+
 ## 0.1.4
 
 ### Patch Changes