+++ packages/test-runner-cli/CHANGELOG.md
@@ -1,5 +1,80 @@
 # @web/test-runner-cli
 
+## 0.6.7
+
+### Patch Changes
+
+- 145a8e6: correctly encode/decode test framework url
+- Updated dependencies [145a8e6]
+  - @web/test-runner-core@0.8.7
+
+## 0.6.6
+
+### Patch Changes
+
+- 304558e: fix(test-runner): deduplicated browsers when reporting
+- Updated dependencies [304558e]
+  - @web/test-runner-core@0.8.6
+
+## 0.6.5
+
+### Patch Changes
+
+- cd8928b: separate reporting per browser launcher
+- Updated dependencies [4edf123]
+- Updated dependencies [cd8928b]
+  - @web/test-runner-core@0.8.5
+
+## 0.6.4
+
+### Patch Changes
+
+- 416c0d2: Update dependencies
+- aadf0fe: Speed up test loading by inling test config and preloading test files.
+- Updated dependencies [416c0d2]
+- Updated dependencies [aadf0fe]
+  - @web/test-runner-chrome@0.7.2
+  - @web/test-runner-core@0.8.4
+
+## 0.6.3
+
+### Patch Changes
+
+- c256a08: allow configuring concurrency per browser launcher
+- Updated dependencies [c256a08]
+  - @web/test-runner-chrome@0.7.1
+  - @web/test-runner-core@0.8.3
+
+## 0.6.2
+
+### Patch Changes
+
+- 0b5cc82: always print stack traces in errors
+
+## 0.6.1
+
+### Patch Changes
+
+- 7ec6e94: don't require files option when using groups
+- 438176c: Allow specifying default test group
+- Updated dependencies [175b124]
+  - @web/test-runner-core@0.8.1
+
+## 0.6.0
+
+### Minor Changes
+
+- 2291ca1: replaced HTTP with websocket for server-browser communication
+
+  this improves test speed, especially when a test file makes a lot of concurrent requests
+  it lets us us catch more errors during test execution, and makes us catch them faster
+
+### Patch Changes
+
+- Updated dependencies [2291ca1]
+  - @web/test-runner-chrome@0.7.0
+  - @web/test-runner-core@0.8.0
+
 ## 0.5.18
 
 ### Patch Changes