+++ packages/test-runner-puppeteer/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @web/test-runner-puppeteer
 
+## 0.7.2
+
+### Patch Changes
+
+- 416c0d2: Update dependencies
+- Updated dependencies [416c0d2]
+- Updated dependencies [aadf0fe]
+  - @web/test-runner-chrome@0.7.2
+  - @web/test-runner-core@0.8.4
+
+## 0.7.1
+
+### Patch Changes
+
+- c256a08: allow configuring concurrency per browser launcher
+- Updated dependencies [c256a08]
+  - @web/test-runner-chrome@0.7.1
+  - @web/test-runner-core@0.8.3
+
+## 0.7.0
+
+### Minor Changes
+
+- 2291ca1: replaced HTTP with websocket for server-browser communication
+
+  this improves test speed, especially when a test file makes a lot of concurrent requests
+  it lets us us catch more errors during test execution, and makes us catch them faster
+
+### Patch Changes
+
+- Updated dependencies [2291ca1]
+  - @web/test-runner-chrome@0.7.0
+  - @web/test-runner-core@0.8.0
+
 ## 0.6.4
 
 ### Patch Changes